requirements.txt
anyio==4.4.0
astor==0.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.11
decorator==5.1.1
filelock==3.16.1
fire==0.6.0
fonttools==4.53.1
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.0
idna==3.10
imageio==2.35.1
imgaug==0.4.0
Jinja2==3.1.4
kiwisolver==1.4.7
lazy_loader==0.4
lmdb==1.5.1
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
paddleocr==2.8.1
paddlepaddle==2.6.2
pandas==2.2.2
pillow==10.4.0
protobuf==5.28.1
psutil==6.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post5
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
PyYAML==6.0.2
rapidfuzz==3.9.7
requests==2.32.3
scikit-image==0.24.0
scipy==1.14.1
seaborn==0.13.2
setuptools==75.1.0
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sympy==1.13.2
termcolor==2.4.0
thop==0.1.1.post2209072238
tifffile==2024.8.30
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
-e git+https://github.com/THU-MIG/yolov10.git@cd2f79c70299c9041fb6d19617ef1296f47575b1#egg=ultralytics
ultralytics-thop==2.0.6
urllib3==2.2.3
